Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main(){
-   int n,m,p;
-   scanf("%d %d %d",&n,&m,&p);
-   int average=(n+m+p)/3;
-   printf("Average: %.2d",&average);
+   float n,m,p;
+   scanf("%f %f %f",&n,&m,&p);
+   float average=(n+m+p)/3;
+   printf("Average: %.2f",&average);
    return 0;
 }
